feat: added get  order status and cancel order and completed place order functionality

+++ backend/src/models/order.rs
@@ -1,3 +1,6 @@
+use sqlx::types::BigDecimal;
+use strum::Display;
+use strum::EnumString;
 use uuid::Uuid;
 
 #[allow(dead_code)]
@@ -5,18 +8,21 @@ pub struct Order {
     pub order_id: Uuid,
     pub user_id: Uuid,
     pub symbol: String,
-    pub quantity: f64,
+    pub quantity: BigDecimal,
     pub price: f64,
     pub order_type: OrderType, // Buy or Sell
     pub status: OrderStatus,
 }
 #[allow(dead_code)]
+#[derive(Debug, Clone, Display, EnumString)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
 pub enum OrderType {
     Buy,
     Sell,
 }
 
-#[allow(dead_code)]
+#[derive(Debug, Clone, Display, EnumString)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
 pub enum OrderStatus {
     Pending,
     Reserved,
